+++ changelog.md
@@ -11,6 +11,7 @@ and this project will adhere to [Semantic Versioning](https://semver.org/spec/v2
 
 - **BREAKING `remirror/react-utils`:** Rename `childIsFunction` to `propIsFunction` and make it a _pseudo_ predicate function (returns true when it doesn't throw an error).
 - Add support for [Git Large File Storage (LFS)](https://git-lfs.github.com/)
+- `remirror/editor-twitter`, `remirror/editor-wysiwyg` : Use image-snapshot testing to ensure SSR and DOM rendered editors are identical
 
 ## [0.2.0] - 2019-06-18